add NodeBase
